Changellogs: re-moved the menus to a ratVentureMenus.py folder.
